Synthesizer/Ableton/requirements.txt
configparser==3.5.0
mido==1.2.8
numpy==1.14.2
pyOSC==0.3.5b5294
